refactor(playground): split project manager module
