+++ packages/agent-playground/src/lib/playground/project-manager/constants.ts
@@ -1,0 +1,11 @@
+const DAYS_PER_WEEK = 7;
+const HOURS_PER_DAY = 24;
+const MINUTES_PER_HOUR = 60;
+const SECONDS_PER_MINUTE = 60;
+const MS_PER_SECOND = 1000;
+
+export const CURRENT_VERSION = '1.0.0';
+export const RANDOM_ID_BASE = 36;
+export const RECENT_ACTIVITY_WINDOW_MS =
+  DAYS_PER_WEEK * HOURS_PER_DAY * MINUTES_PER_HOUR * SECONDS_PER_MINUTE * MS_PER_SECOND;
+export const STORAGE_KEY = 'robota-playground-projects';
