+++ packages/agent-playground/src/lib/playground/project-manager/index.ts
@@ -1,0 +1,9 @@
+export { ProjectManager } from './project-manager';
+export type {
+  IPlaygroundProject,
+  IProjectConfig,
+  IProjectMetadata,
+  IProjectSettings,
+  IProjectStats,
+  TPlaygroundProvider,
+} from './types';
